ページコメントのセグメント作成機能を追加し、通知作成機能を実装しました。また、関連するインポートパスを修正し、不要なファイルを削除しました。
