+++ next/src/app/[locale]/(common-layout)/user/[handle]/page/[slug]/_components/comment/_components/page-comment-form/_lib/process-page-comment-html.ts
@@ -10,13 +10,13 @@ import remarkGfm from "remark-gfm";
 import remarkRehype from "remark-rehype";
 import { unified } from "unified";
 import type { Plugin } from "unified";
-import { upsertPageComment } from "../../../_db/mutations.server";
-import { createPageCommentSegments } from "../../../_db/mutations.server";
+import { upsertPageComment } from "../_db/mutations.server";
+import { createPageCommentSegments } from "../_db/mutations.server";
 export function parseHtmlToAst(html: string): Root {
 	return unified().use(rehypeParse, { fragment: true }).parse(html) as Root;
 }
 
-export function rehypeAddDataId(commentId: number): Plugin<[], Root> {
+function rehypeAddDataId(commentId: number): Plugin<[], Root> {
 	return function attacher() {
 		return async (tree: Root) => {
 			const blocks = collectBlocksFromRoot(tree);
